+++ modules/common/CHANGELOG.md
@@ -17,6 +17,8 @@ TEMPLATE:
 -->
 ## [UPCOMING]
 - Adds new error `EnsureAllowanceError`
+- Adds new error `InvalidPrecisionError`
+
 ## 0.9.1-alpha
 On 2022-12-13 15:57:38
 ### Fixed